requirements.txt
torch
torchvision
albumentations
tensorboard
scikit-learn
matplotlib
seaborn
pyyaml